ports: add lsd-bin (#78)
* ports: add lsd-bin

* Update pkgfile

small cleanup

Also to anyone reading, no this is NOT from the bonsai website, this is NOT an official binary. To compile this (or exa) requires rust which is a huge build that will be far from now. A user heavily wished we would add this so this is us obliging.